Rename crate
