+++ core/db/src/error.rs
