+++ core/rust_sim/examples/forked_sim/main.rs
@@ -3,7 +3,7 @@ mod dai_abi;
 use dai_abi::ABI as DAI_ABI;
 
 use alloy_primitives::{Address, U256};
-use fork_evm::ForkDb;
+use db::ForkDb;
 use rust_sim::network::Network;
 use rust_sim::utils::address_from_hex;
