+++ core/rust_sim/src/network/mod.rs
@@ -4,8 +4,8 @@ use crate::utils::Eth;
 use alloy_primitives::{Address, Uint, B256, U256};
 use alloy_sol_types::SolCall;
 use anyhow::Result;
+use db::{ForkDb, LocalDB, Requests, DB};
 pub use ethereum_types::U64;
-use fork_evm::{ForkDb, LocalDB, Requests, DB};
 use log::debug;
 use revm::primitives::{AccountInfo, Bytecode, ExecutionResult, Log, ResultAndState, TxEnv};
 use revm::EVM;
