+++ rust/Cargo.toml
@@ -19,4 +19,4 @@ pyo3 = { version="0.20.0", features = ["extension-module"] }
 pyo3-log = "0.9.0"
 
 rust_sim = { path = "../core/rust_sim" }
-fork-evm = { path = "../core/fork_evm" }
+db = { path = "../core/db" }
