+++ rust/src/sim/empty_env.rs
@@ -3,7 +3,7 @@ use super::snapshot::PyDbState;
 use super::base_env::BaseEnv;
 use super::snapshot;
 use crate::types::{PyAddress, PyEvent, PyExecutionResult, PyRevertError};
-use fork_evm::LocalDB;
+use db::LocalDB;
 use pyo3::exceptions::PyRuntimeError;
 use pyo3::prelude::*;
 use pyo3::types::PyBytes;
