+++ rust/src/sim/snapshot.rs
@@ -1,5 +1,5 @@
 use alloy_primitives::{Address, Bytes, B256, U256};
-use fork_evm::{LocalDB, Requests, DB};
+use db::{LocalDB, Requests, DB};
 use pyo3::{types::PyBytes, Python};
 use revm::{
     db::{AccountState, DbAccount},
